Only redirect on successful message receipt
If you forgot to fill in some information in the contact form you'd be shown the success page with a bunch of errors on top. Most likely you'd miss the error message and mistakenly think your contact has gone through. Less likely you'd try to go back to the previous page, but your contact message would be gone. This change solves both issues.